Add UsingRhsConst and UsingLhsConst variations to shift tril tests
Previously, the tril tests for the shift opcodes only tested these
opcodes with both operands as constants and with both operands as loads.
However, the codegens have many special cases that are used only when
one of the two operands is a constant. To achieve better test coverage,
new variants of the shift opcode tests have been added which use a
constant for one of the operands and a load for the other.

Signed-off-by: Ben Thomas <ben@benthomas.ca>